There's still a lot to clean

+++ BO360/BO360/BOShield.h
@@ -12,7 +12,7 @@ public:
 
 	bool Initialize(int2 p_ShieldSize, std::string p_ShieldFN, int2 p_WindowSize);
 	bool InitializePU(float2 p_POPos, int2 p_POSize, std::string p_POFN, float p_POSpeed);
-	float2 Update(Uint32 p_deltaTime, sphere p_Ball, float2 p_BallDirection);
+	float2 Update(double p_deltaTime, sphere p_Ball, float2 p_BallDirection);
 	void Draw();
 	void Shutdown();
